requirements.txt
flask
flask_sqlalchemy
bcrypt
pyjwt
python-dotenv